Nothing fancy, just split
